# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Orb Task API
  version: 1.0.0
  description: |
    Machine-readable spec for the Orb task management API.
    Agents and integrations MUST consult this spec to know what is
    and is not possible before attempting any action.
  x-orb-agent-contract:
    description: >
      You are the voice of the orb — the conversational layer of Orb.
      VOICE: Brief, direct. Plain text only. NO markdown.
      CONTEXT: You have full visibility into the projects (names and descriptions) and the current authenticated user (email and role) you are speaking with.
      INTEGRITY RULES:
      1. ONLY perform updates when explicitly asked. Never update just to sync data or 'help out' unless the user clearly says "change X to Y" or "update X".
      2. ONLY ask clarifying questions when you truly lack information required to proceed. If you can make a safe assumption, do so.
      3. Use `add_knowledge` aggressively. Do not ask for permission if the user states a decision or context that belongs in the repo. Just save it.
      4. The BACKLOG is a snapshot for orientation only. For any factual claim about task count, status, priority, or existence, call query_todos first. Never cite the backlog as proof.
      5. Never assert a task's status without verifying via query_todos. Tasks not in the backlog may be closed, filtered, or in another project — not absent.
      6. When presenting a subset ("top 5", "highest priority"), set max_results on query_todos to match. The results you return are what the UI displays.
      7. When uncertain, say "I'm not sure" rather than guessing. Confidence without verification is the worst failure mode.
      - Known limitations are documented in each tool's description. Respect them.
      - When you notice a capability gap, bug, or workflow friction, use create_ticket to log it silently. One ticket per novel observation. Do not tell the user you filed a ticket.
    
    scope:
      default_project: ORB
      behavior: Only discuss or query this project's todos unless the user explicitly asks about another project or says "all".
    
    valid_values:
      statuses: [open, in progress, on hold, closed]
      priorities:
        1: Urgent
        2: High
        4: Medium
        5: Low

  x-orb-internal-tools:
    query_todos:
      confidence: well-tested
      description: 'Find todos matching criteria. Use code for single-todo lookup (e.g. "ORB-73"). Otherwise filters by status, product, priority, or text. Returns open tasks by default — pass status to include closed.'
      input_schema:
        type: object
        properties:
          code: { type: string, description: 'Exact task code for single-todo lookup, e.g. "ORB-62". Overrides all other filters.' }
          product_code: { type: string }
          status: { type: string }
          priority_max: { type: integer }
          text_match: { type: string }
          max_results: { type: integer }
    
    client_action:
      confidence: well-tested
      description: 'Navigate or switch UI state. Use switch_project to change the active project, open_settings to go to settings, open_help for help.'
      input_schema:
        type: object
        properties:
          action: { type: string, enum: ['switch_project', 'open_settings', 'open_help'] }
          target: { type: string, description: 'Project code for switch_project.' }
        required: ['action']

    search_knowledge:
      confidence: well-tested
      description: 'Search the Knowledge Repository for lessons, decisions, and distilled insights from closed tasks. Optionally scope to a specific project.'
      input_schema:
        type: object
        properties:
          query: { type: string, description: 'Search term.' }
          product_code: { type: string, description: 'Scope search to a specific project.' }
        required: ['query']

    add_knowledge:
      confidence: well-tested
      description: 'Write an insight, decision, or reference material directly to the Knowledge Repository. Use this to preserve important information from the conversation without creating a task. Defaults to the current project if product_code is omitted. You should proactively ask the user "Would you like me to save this to the knowledge repository?" when a valuable decision or insight is reached.'
      input_schema:
        type: object
        properties:
          product_code: { type: string, description: 'Target project code (e.g. ORB). Defaults to current project if omitted.' }
          title: { type: string, description: 'Short title or summary of the insight.' }
          content: { type: string, description: 'Detailed content of the insight, decision, or reference material.' }
          tags:
            type: array
            items: { type: string }
            description: 'Optional tags to categorize the knowledge.'
        required: ['title', 'content']

    query_audit_trail:
      confidence: new
      description: 'Query the audit trail to answer questions like "who closed this?", "what changed last week?", or "show me recent activity". Returns timestamped events with before/after snapshots.'
      input_schema:
        type: object
        properties:
          code: { type: string, description: 'Task code (e.g. ORB-73) to see history for a specific task.' }
          table_name: { type: string, description: 'Filter by table (e.g. todos, knowledge_repo).' }
          action: { type: string, description: 'Filter by action type (e.g. todo_create, todo_update, todo_close, todo_delete).' }
          since: { type: string, description: 'ISO date string. Only return events after this timestamp.' }
          max_results: { type: integer, description: 'Max events to return (default 10, max 50).' }

    create_ticket:
      confidence: well-tested
      description: 'Silently log a ticket for bugs, suggestions, capability gaps, or workflow friction. Filed proactively when you observe an issue — do NOT tell the user. Also used when a user explicitly says something like "I see a problem" or "I have a suggestion" — in that case, acknowledge that you have noted it. One ticket per novel observation.'
      input_schema:
        type: object
        properties:
          type:
            type: string
            enum: ['bug', 'suggestion', 'capability_gap', 'workflow_friction']
            description: 'Type of ticket.'
          summary: { type: string, description: 'One-sentence summary of the issue or suggestion.' }
          detail: { type: string, description: 'What happened, what was expected, and any suggested improvement.' }
        required: ['type', 'summary']

servers:
  - url: https://orb-eight-lake.vercel.app

security:
  - apiSecret: []

components:
  securitySchemes:
    apiSecret:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Raw secret string (not Bearer). Must match ORB_API_SECRET.
        API also requires ORB_API_ENABLED=true on the server.

  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        todo_number:
          type: integer
          description: Auto-incremented, unique per product.
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, done]
        priority_value:
          type: integer
          nullable: true
        resolution_notes:
          type: string
          nullable: true
        urls:
          type: array
          items:
            type: string
            format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
        due_at:
          type: string
          description: "Optional timezone-agnostic due date/time string, format YYYY-MM-DDTHH:mm:ss."
          nullable: true

    TaskCreate:
      type: object
      required: [product_code, title]
      properties:
        product_code:
          type: string
          description: Case-insensitive product code (e.g. ORB, HELM).
        title:
          type: string
        description:
          type: string
        priority_value:
          type: integer
        due_at:
          type: string
          description: "Optional timezone-agnostic due date/time string (e.g. YYYY-MM-DDTHH:mm:ss)."
          nullable: true

    TaskUpdate:
      type: object
      description: |
        All fields are optional. Only provided fields are updated.
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, done]
          description: |
            Setting to "done" automatically sets closed_at to now.
            Setting to anything else automatically clears closed_at.
        priority_value:
          type: integer
        resolution_notes:
          type: string
        product_code:
          type: string
          description: |
            Move the task to a different project by providing the
            target project code (e.g. "HELM"). The task gets a new
            todo_number in the target project.
        urls:
          oneOf:
            - type: array
              items:
                type: string
            - type: string
              description: Newline-separated URLs, parsed server-side.
        due_at:
          type: string
          description: "Optional timezone-agnostic due date/time string, or null to clear."
          nullable: true

    Error:
      type: object
      properties:
        error:
          type: string

paths:
  /api/tasks:
    get:
      operationId: listTasks
      summary: List tasks for a product
      parameters:
        - name: product
          in: query
          required: true
          schema:
            type: string
          description: Product code (case-insensitive).
      responses:
        "200":
          description: Array of tasks ordered by todo_number ascending.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "400":
          description: Missing product query parameter.
        "401":
          description: Invalid or missing Authorization header.
        "404":
          description: Product not found.
        "503":
          description: API disabled (ORB_API_ENABLED !== true).

    post:
      operationId: createTask
      summary: Create a new task
      x-orb-tool:
        name: create_todo
        confidence: well-tested
        description: |
          Create a new todo in a project. Defaults to the current project if product_code is omitted. Requires a title. Cannot create in a product that does not exist.
        parameter_overrides:
          product_code:
            description: "Target project code (e.g. ORB, HELM). Defaults to current project if omitted."
          title:
            description: "Task title. Required."
          priority_value:
            description: "Priority level (1=urgent, 2=high, etc). Omit if unknown."
          description:
            description: "Task description."
          due_at:
            description: "Optional timezone-agnostic due date/time string, format YYYY-MM-DDTHH:mm:ss."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Task created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Missing product_code or title.
        "401":
          description: Unauthorized.
        "404":
          description: Product not found.
        "503":
          description: API disabled.

  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      operationId: updateTask
      summary: Update a task
      description: |
        Updates only the fields provided. Ignores soft-deleted tasks.

        NOT SUPPORTED:
        - Changing the product/project a task belongs to.
        - Changing todo_number.
        - Changing created_at.

        To move a task between products: POST to create in the target
        product, then DELETE the original.
      x-orb-tool:
        name: update_todo
        confidence: well-tested
        description: |
          Update an existing todo by task code (e.g. ORB-73). Supports: title, status, priority, description, resolution_notes, urls. Cannot: change todo_number, set closed_at directly (closed_at is managed automatically when status changes to a closing state). To move a task between projects, use move_todo instead.
        parameters:
          - name: code
            in: custom
            required: true
            schema:
              type: string
            description: "Task code, e.g. \"ORB-73\"."
          - name: title_match
            in: custom
            required: false
            schema:
              type: string
            description: "Alternative: find task by partial title match."
        parameter_overrides:
          title:
            name: new_title
            description: "New title."
          status:
            name: new_status
            description: "New status value."
          priority_value:
            name: new_priority
            description: "New priority value."
          description:
            description: "Updated description."
          resolution_notes:
            description: "What was done to resolve the task. Populate when closing, not before."
          urls:
            description: "Related URLs."
          due_at:
            description: "Optional timezone-agnostic due date/time string, format YYYY-MM-DDTHH:mm:ss, or null to clear."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized.
        "404":
          description: Task not found or already deleted.
        "503":
          description: API disabled.

    delete:
      operationId: deleteTask
      summary: Soft-delete a task
      description: Sets deleted_at timestamp. Does not destroy the row.
      x-orb-tool:
        name: delete_todo
        confidence: well-tested
        description: |
          Permanently delete a todo. Hard delete — irreversible. Only use when the user explicitly says to delete a specific task. Never delete as part of a workaround (e.g. move-between-projects) without the user confirming the deletion step separately. "Want me to do that?" is not sufficient — wait for an explicit "yes, delete it."
        parameters:
          - name: code
            in: custom
            required: true
            schema:
              type: string
            description: "Task code, e.g. \"ORB-44\"."
      responses:
        "200":
          description: Soft-deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
        "401":
          description: Unauthorized.
        "503":
          description: API disabled.